revert findfile to support/init.py and fix pep8 style
